Extract module root
