+++ scripts/update_templates/files.go
@@ -41,22 +41,3 @@ func openGitHubSources() (*os.File, error) {
 
 	return os.Open(filepath.Join(repoRoot, filepath.FromSlash(relativeSourcesPath)))
 }
-
-func findModuleRoot() (string, error) {
-	dir, err := os.Getwd()
-	if err != nil {
-		return "", err
-	}
-
-	for {
-		if _, err := os.Stat(filepath.Join(dir, "go.mod")); err == nil {
-			return dir, nil
-		}
-
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			return "", os.ErrNotExist
-		}
-		dir = parent
-	}
-}
